requirements.txt
django
pillow
torch
torchvision
